+++ dev/bench_imread2.py
@@ -4,43 +4,39 @@ def imread2_bench():
     """
     Another imread benchmark
     """
-    import itertools as it
-
-    def basis_product(basis):
-        """
-        Generates the Cartesian product of the ``basis.values()``, where each
-        generated item labeled by ``basis.keys()``.
-        """
-        keys = list(basis.keys())
-        for vals in it.product(*basis.values()):
-            kw = dict(zip(keys, vals))
-            yield kw
+    import ubelt as ub
 
     data_basis = {
         'dsize': [
             (32, 32),
             (128, 128),
             (512, 512),
-            (800, 800),
-            (1200, 1200)
+            # (800, 800),
+            # (1200, 1200)
         ],
         'channels': [
             3,
-            10,
+            13,
             64,
-            128,
+            # 128,
         ],
         'dtype': [
             'uint8',
             'float32'
         ],
     }
-    data_profiles = list(basis_product(data_basis))
+    data_profiles = list(ub.named_product(data_basis))
 
-    method_profiles = [
-        # {'backend': 'skimage'},
-        # {'backend': 'gdal'},
-        {'backend': 'cv2', 'space': None},
+    imread_profiles = [
+        {'backend': 'skimage'},
+        {'backend': 'gdal'},
+        # {'backend': 'cv2', 'space': None},
+    ]
+
+    imwrite_profiles = [
+        {'backend': 'skimage'},
+        {'backend': 'gdal'},
+        # {'backend': 'cv2', 'space': None},
     ]
 
     import numpy as np
@@ -51,36 +47,48 @@ def imread2_bench():
     dpath = ub.ensure_app_cache_dir('kwimage/bench/imread2')
 
     import timerit
-    read_ti = timerit.Timerit(2, bestof=1, verbose=2)
+    read_ti = timerit.Timerit(1, bestof=1, verbose=2)
 
     measures = []
     for data_profile in data_profiles:
-        data_profile_key = ub.repr2(data_profile, nobr=1, itemsep='', keysep='=', nl=0, sv=1, sk=1, kvsep='=', sort=0)
+        data_profile_key = ub.repr2(data_profile, compact=1, sort=0)
         width, height = data_profile['dsize']
         channels = data_profile['channels']
         shape = (height, width, channels)
 
         data = (np.random.rand(*shape) * 255).astype(data_profile['dtype'])
         fpath = join(dpath, data_profile_key + '.tiff')
-        kwimage.imwrite(fpath, data, backend='gdal')
 
-        for method_profile in method_profiles:
+        for imwrite_profile in imwrite_profiles:
 
-            if method_profile['backend'] == 'cv2' and channels > 4:
-                continue
+            imwrite_profile_key = ub.repr2(imwrite_profile, compact=1, sort=0)
+            kwimage.imwrite(fpath, data, **imwrite_profile)
 
-            method_profile_key = ub.repr2(method_profile, nobr=1, itemsep='', keysep='=', nl=0, sv=1, sk=1, kvsep='=', sort=0)
-            for timer in read_ti.reset(f'{data_profile_key} {method_profile_key}'):
-                with timer:
-                    kwimage.imread(fpath, **method_profile)
+            for imread_profile in imread_profiles:
 
-            time_result = {
-                'mean': read_ti.mean(),
-                'std': read_ti.std(),
-                'min': read_ti.min(),
-            }
-            row = ub.dict_union(data_profile, method_profile, time_result)
-            measures.append(row)
+                if imread_profile['backend'] == 'cv2' and channels > 4:
+                    continue
+
+                imread_profile_key = ub.repr2(imread_profile, compact=1, sort=0)
+
+                io_profile_key = 'imwrite({})->imread({})'.format(
+                    imwrite_profile_key, imread_profile_key)
+
+                for timer in read_ti.reset(f'{data_profile_key} {io_profile_key}'):
+                    with timer:
+                        kwimage.imread(fpath, **imread_profile)
+
+                time_result = {
+                    'mean': read_ti.mean(),
+                    'std': read_ti.std(),
+                    'min': read_ti.min(),
+                }
+                row = ub.dict_union(data_profile, time_result)
+                row['imwrite_profile_key'] = imwrite_profile_key
+                row['imread_profile_key'] = imread_profile_key
+                row['data_profile_key'] = data_profile_key
+                row['io_profile_key'] = io_profile_key
+                measures.append(row)
 
     import pandas as pd
     for row in measures:
@@ -99,7 +107,23 @@ def imread2_bench():
 
     import kwplot
     sns = kwplot.autosns()
+    plt = kwplot.autoplt()
 
-    ax = sns.lineplot(data=df, x='bits', y='min', hue='backend')
+    df['io_profile_key'] = (
+        'imwrite(' + df['imwrite_profile_key'] + ')' +
+        '->' +
+        'imread(' + df['imread_profile_key'] + ')'
+    )
+    ax = sns.lineplot(data=df, x='bits', y='min', hue='io_profile_key')
     ax.set_xscale('log')
     ax.set_yscale('log')
+
+    plt.show()
+
+
+if __name__ == '__main__':
+    """
+    CommandLine:
+        python ~/code/kwimage/dev/bench_imread2.py
+    """
+    imread2_bench()