+++ kwimage/im_alphablend.py
@@ -26,7 +26,7 @@ def overlay_alpha_layers(layers, keepalpha=True, dtype=np.float32):
         >>> keys = ['astro', 'carl', 'stars']
         >>> layers = [kwimage.grab_test_image(k, dsize=(100, 100)) for k in keys]
         >>> layers = [kwimage.ensure_alpha_channel(g, alpha=.5) for g in layers]
-        >>> stacked = overlay_alpha_layers(layers)
+        >>> stacked = kwimage.overlay_alpha_layers(layers)
         >>> # xdoctest: +REQUIRES(--show)
         >>> import kwplot
         >>> kwplot.autompl()
@@ -76,12 +76,37 @@ def overlay_alpha_images(img1, img2, keepalpha=True, dtype=np.float32,
         >>> img1 = kwimage.grab_test_image('astro', dsize=(100, 100))
         >>> img2 = kwimage.grab_test_image('carl', dsize=(100, 100))
         >>> img1 = kwimage.ensure_alpha_channel(img1, alpha=.5)
-        >>> img3 = overlay_alpha_images(img1, img2)
+        >>> img3 = kwimage.overlay_alpha_images(img1, img2)
         >>> # xdoctest: +REQUIRES(--show)
         >>> import kwplot
         >>> kwplot.autompl()
         >>> kwplot.imshow(img3)
         >>> kwplot.show_if_requested()
+
+    Ignore:
+        import numpy as np
+        import kwimage
+        poly = kwimage.Polygon.random().scale((10, 10))
+
+        img2 = np.zeros((10, 10, 4))
+
+        img1 = np.zeros((10, 10))
+        indicator = poly.fill(img1)
+
+        to_overlay = np.zeros((10, 10) + (4,), dtype=np.float32)
+        to_overlay = kwimage.Mask(indicator, format='c_mask').draw_on(to_overlay, color='lime')
+        to_overlay = kwimage.ensure_alpha_channel(to_overlay)
+        to_overlay[..., 3] = (indicator > 0).astype(np.float32) * 0.5
+
+        raster = kwimage.overlay_alpha_images(to_overlay, img2)
+
+        # xdoctest: +REQUIRES(--show)
+        import kwplot
+        kwplot.autompl()
+        kwplot.imshow(raster)
+        kwplot.show_if_requested()
+
+
     """
     rgb1, alpha1 = _prep_rgb_alpha(img1, dtype=dtype)
     rgb2, alpha2 = _prep_rgb_alpha(img2, dtype=dtype)