+++ kwimage/im_cv2.py
@@ -22,6 +22,7 @@ _CV2_INTERPOLATION_TYPES = {
 }
 
 
+# https://docs.opencv.org/3.4/d2/de8/group__core__array.html#ga209f2f4869e304c82d07739337eae7c5
 _CV2_BORDER_MODES = {
     'constant':    cv2.BORDER_CONSTANT,
     'replicate':   cv2.BORDER_REPLICATE,
@@ -29,6 +30,7 @@ _CV2_BORDER_MODES = {
     'wrap':        cv2.BORDER_WRAP,
     'reflect101':  cv2.BORDER_REFLECT101,
     'transparent': cv2.BORDER_TRANSPARENT,
+    # 'isolated':    cv2.BORDER_ISOLATED,
 }
 
 
@@ -787,20 +789,189 @@ def gaussian_patch(shape=(7, 7), sigma=None):
         >>> kwplot.show_if_requested()
     """
     if sigma is None:
-        sigma1 = 0.3 * ((shape[0] - 1) * 0.5 - 1) + 0.8
-        sigma2 = 0.3 * ((shape[1] - 1) * 0.5 - 1) + 0.8
+        sigma_x = 0.3 * ((shape[0] - 1) * 0.5 - 1) + 0.8
+        sigma_y = 0.3 * ((shape[1] - 1) * 0.5 - 1) + 0.8
     elif isinstance(sigma, (float, int)):
-        sigma1 = sigma2 = sigma
+        sigma_x = sigma_y = sigma
     else:
-        sigma1, sigma2 = sigma
+        sigma_x, sigma_y = sigma
     # see hesaff/src/helpers.cpp : computeCircularGaussMask
-    kernel_d0 = cv2.getGaussianKernel(shape[0], sigma1)
-    if shape[0] == shape[1] and sigma2 == sigma1:
+    kernel_d0 = cv2.getGaussianKernel(shape[0], sigma_x)
+    if shape[0] == shape[1] and sigma_y == sigma_x:
         kernel_d1 = kernel_d0
     else:
-        kernel_d1 = cv2.getGaussianKernel(shape[1], sigma2)
+        kernel_d1 = cv2.getGaussianKernel(shape[1], sigma_y)
     gausspatch = kernel_d0.dot(kernel_d1.T)
     return gausspatch
+
+
+def _auto_kernel_sigma(kernel=None, sigma=None, autokernel_mode='ours'):
+    """
+    Attempt to determine sigma and kernel size from heuristics
+
+    Example:
+        >>> from kwimage.im_cv2 import *  # NOQA
+        >>> _auto_kernel_sigma(None, None)
+        >>> _auto_kernel_sigma(3, None)
+        >>> _auto_kernel_sigma(None, 0.8)
+        >>> _auto_kernel_sigma(7, None)
+        >>> _auto_kernel_sigma(None, 1.4)
+
+    Ignore:
+        >>> # xdoctest: +REQUIRES(--demo)
+        >>> from kwimage.im_cv2 import *  # NOQA
+        >>> rows = []
+        >>> for k in np.arange(3, 101, 2):
+        >>>     s = _auto_kernel_sigma(k, None)[1][0]
+        >>>     rows.append({'k': k, 's': s, 'type': 'auto_sigma'})
+        >>> #
+        >>> sigmas = np.array([r['s'] for r in rows])
+        >>> other = np.linspace(0, sigmas.max() + 1, 100)
+        >>> sigmas = np.unique(np.hstack([sigmas, other]))
+        >>> sigmas.sort()
+        >>> for s in sigmas:
+        >>>     k = _auto_kernel_sigma(None, s, autokernel_mode='cv2')[0][0]
+        >>>     rows.append({'k': k, 's': s, 'type': 'auto_kernel (cv2)'})
+        >>> #
+        >>> for s in sigmas:
+        >>>     k = _auto_kernel_sigma(None, s, autokernel_mode='ours')[0][0]
+        >>>     rows.append({'k': k, 's': s, 'type': 'auto_kernel (ours)'})
+        >>> import pandas as pd
+        >>> df = pd.DataFrame(rows)
+        >>> p = df.pivot(['s'], ['type'], ['k'])
+        >>> print(p[~p.droplevel(0, axis=1).auto_sigma.isnull()])
+        >>> # xdoctest: +REQUIRES(--show)
+        >>> import kwplot
+        >>> sns = kwplot.autosns()
+        >>> sns.lineplot(data=df, x='s', y='k', hue='type')
+    """
+    import numbers
+    if kernel is None and sigma is None:
+        kernel = 3
+
+    if kernel is not None:
+        if isinstance(kernel, numbers.Integral):
+            k_x = k_y = kernel
+        else:
+            k_x, k_y = kernel
+
+    if sigma is None:
+        # https://github.com/egonSchiele/OpenCV/blob/09bab41/modules/imgproc/src/smooth.cpp#L344
+        sigma_x = 0.3 * ((k_x - 1) * 0.5 - 1) + 0.8
+        sigma_y = 0.3 * ((k_y - 1) * 0.5 - 1) + 0.8
+    else:
+        if isinstance(sigma, numbers.Number):
+            sigma_x = sigma_y = sigma
+        else:
+            sigma_x, sigma_y = sigma
+
+    if kernel is None:
+        USE_CV2_DEF = 0
+        if autokernel_mode == 'zero':
+            # When 0 computed internally via cv2
+            k_x = k_y = 0
+        elif autokernel_mode == 'cv2':
+            # if USE_CV2_DEF:
+            # This is the CV2 definition
+            # https://github.com/egonSchiele/OpenCV/blob/09bab41/modules/imgproc/src/smooth.cpp#L387
+            depth_factor = 3  # or 4 for non-uint8
+            k_x = int(round(sigma_x * depth_factor * 2 + 1)) | 1
+            k_y = int(round(sigma_y * depth_factor * 2 + 1)) | 1
+        elif autokernel_mode == 'ours':
+            # But I think this definition makes more sense because it keeps
+            # sigma and the kernel in agreement more often
+            """
+            # Our hueristic is computed via solving the sigma heuristic for k
+            import sympy as sym
+            s, k = sym.symbols('s, k', rational=True)
+            sa = sym.Rational('3 / 10') * ((k - 1) / 2 - 1) + sym.Rational('8 / 10')
+            sym.solve(sym.Eq(s, sa), k)
+            """
+            k_x = max(3, round(20 * sigma_x / 3 - 7/3)) | 1
+            k_y = max(3, round(20 * sigma_y / 3 - 7/3)) | 1
+        else:
+            raise KeyError(autokernel_mode)
+    sigma = (sigma_x, sigma_y)
+    kernel = (k_x, k_y)
+    return kernel, sigma
+
+
+def gaussian_blur(image, kernel=None, sigma=None, border_mode=None, dst=None):
+    """
+    Apply a gausian blur to an image.
+
+    This is a simple wrapper around :func:`cv2.GaussianBlur` with concise
+    parametarization and sane defaults.
+
+    Args:
+        image (ndarray):
+            the input image
+
+        kernel (int | Tuple[int, int]):
+            The kernel size in x and y directions.
+
+        sigma (float | Tuple[float, float]):
+            The gaussian spread in x and y directions.
+
+        border_mode (str | int | None):
+            Border text code or cv2 integer. Border codes are 'constant'
+            (default), 'replicate', 'reflect', 'reflect101', and 'transparent'.
+
+        dst (ndarray | None): optional inplace-output array.
+
+    Returns:
+        ndarray: the blurred image
+
+    Example:
+        >>> import kwimage
+        >>> image = kwimage.ensure_float01(kwimage.grab_test_image('astro'))
+        >>> blurred1 = kwimage.gaussian_blur(image)
+        >>> blurred2 = kwimage.gaussian_blur(image, kernel=9)
+        >>> blurred3 = kwimage.gaussian_blur(image, sigma=2)
+        >>> blurred4 = kwimage.gaussian_blur(image, sigma=(2, 5), kernel=5)
+        >>> # xdoc: +REQUIRES(--show)
+        >>> import kwplot
+        >>> kwplot.autompl()
+        >>> pnum_ = kwplot.PlotNums(nRows=4, nCols=1)
+        >>> blurs = [blurred1, blurred2, blurred3, blurred4]
+        >>> for blurred in blurs:
+        >>>     diff = np.abs(image - blurred)
+        >>>     stack = kwimage.stack_images([image, blurred, diff], pad=10, axis=1)
+        >>>     kwplot.imshow(stack, pnum=pnum_())
+        >>> kwplot.show_if_requested()
+    """
+    if kernel is None and sigma is None:
+        kernel = 3
+
+    if kernel is not None:
+        if isinstance(kernel, int):
+            k_x = k_y = kernel
+        else:
+            k_x, k_y = kernel
+
+    if sigma is not None:
+        if isinstance(sigma, (float, int)):
+            sigma_x = sigma_y = sigma
+        else:
+            sigma_x, sigma_y = sigma
+
+    if sigma is None:
+        # https://dsp.stackexchange.com/questions/10057/gaussian-blur-standard-deviation-radius-and-kernel-size
+        # sigma_x = 0.3 * ((k_x - 1) * 0.5 - 1) + 0.8
+        # sigma_y = 0.3 * ((k_y - 1) * 0.5 - 1) + 0.8
+        # When 0 computed via cv2 from kernel
+        sigma_x = sigma_y = 0
+
+    if kernel is None:
+        # When 0 computed via cv2 from sigma
+        k_x, k_y = 0, 0
+
+    borderType = _coerce_border(border_mode)
+    blurred = cv2.GaussianBlur(
+        image, (k_x, k_y), sigmaX=sigma_x, sigmaY=sigma_y,
+        borderType=borderType, dst=dst
+    )
+    return blurred
 
 
 def warp_affine(image, transform, dsize=None, antialias=False,
@@ -843,8 +1014,8 @@ def warp_affine(image, transform, dsize=None, antialias=False,
             nearest, cubic, lancsoz, and area.
 
         border_mode (str):
-            Border code or cv2 integer. Border codes are constant replicate,
-            reflect, wrap, reflect101, and transparent.
+            Border code or cv2 integer. Border codes are constant (default)
+            replicate, reflect, wrap, reflect101, and transparent.
 
         border_value (int | float):
             Used as the fill value if border_mode is constant. Otherwise this