+++ kwimage/structs/heatmap.py
@@ -178,6 +178,22 @@ class _HeatmapDrawMixin(object):
             >>> kwplot.autompl()
             >>> kwplot.imshow(colormask1, fnum=1, title='output space')
             >>> kwplot.show_if_requested()
+
+        Ignore:
+            >>> # xdoctest: +REQUIRES(module:kwplot)
+            >>> import kwarray
+            >>> import kwimage
+            >>> rng = kwarray.ensure_rng(0)
+            >>> class_probs = np.zeros((2, 32, 32))
+            >>> class_probs[0] = kwimage.Polygon.random(rng=rng).scale(16).translate(16).fill(class_probs[0], value=0.5)
+            >>> class_probs[1] = kwimage.Polygon.random(rng=rng).scale(32).fill(class_probs[1], value=0.5)
+            >>> self = kwimage.Heatmap(class_probs=class_probs)
+            >>> canvas = self.colorize()
+            >>> canvas = kwimage.overlay_alpha_images(canvas, np.zeros_like(canvas[:, :, 0:3]))
+            >>> # xdoctest: +REQUIRES(--show)
+            >>> import kwplot
+            >>> kwplot.autompl()
+            >>> kwplot.imshow(canvas)
         """
         import kwplot
 
@@ -1262,7 +1278,7 @@ class Heatmap(_generic.Spatial, _HeatmapDrawMixin,
     @classmethod
     def random(cls, dims=(10, 10), classes=3, diameter=True, offset=True,
                keypoints=False, img_dims=None, dets=None, nblips=10, noise=0.0,
-               rng=None):
+               rng=None, ensure_background=True):
         """
         Creates dummy data, suitable for use in tests and benchmarks
 
@@ -1354,8 +1370,9 @@ class Heatmap(_generic.Spatial, _HeatmapDrawMixin,
             dets = kwimage.Detections.random(num=nblips, scale=img_dims,
                                              keypoints=keypoints,
                                              classes=classes, rng=rng)
-            if 'background' not in dets.classes:
-                dets.classes.append('background')
+            if ensure_background:
+                if 'background' not in dets.classes:
+                    dets.classes.append('background')
 
             classes = dets.classes
         else: